requirements.txt
apturl==0.5.2
asgiref==3.6.0
aspose-words==23.3.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
CacheControl==0.12.11
cachetools==5.3.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
deprecation==2.1.0
distro==1.7.0
distro-info===1.1build1
Django==4.1.6
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
docx2pdf==0.1.8
duplicity==0.8.21
fasteners==0.14.1
fire==0.5.0
firebase==4.0.1
firebase-admin==6.1.0
fonttools==4.38.0
fpdf==1.7.2
future==0.18.2
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.20.2
idna==3.3
img2pdf==0.4.4
importlib-metadata==4.6.4
jeepney==0.7.1
jwcrypto==1.4.2
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.6.3
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
msgpack==1.0.4
netifaces==0.11.0
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
paramiko==2.9.3
pdf2docx==0.5.6
pdf2image==1.16.3
pexpect==4.8.0
pikepdf==7.1.1
Pillow==9.0.1
proto-plus==1.22.2
protobuf==4.22.0rc3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.20.1
pycryptodome==3.17
pycups==2.0.1
PyGetWindow==0.0.9
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMuPDF==1.21.1
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
Pyrebase4==4.6.0
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-jwt==4.0.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.10.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
tabula==1.0.5
termcolor==2.2.0
tinify==1.6.0
tqdm==4.65.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
